Api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
mysqlclient==2.2.0